Moved html specific helper methods to it's own module
